extract calibration from system state
